refactor: split networking code in various modules
