+++ ant-node/src/networking/driver/cmd.rs
@@ -7,33 +7,28 @@
 // permissions and limitations relating to use of the SAFE Network Software.
 
 use crate::networking::{
-    driver::{PendingGetClosestType, SwarmDriver},
+    driver::{event::MsgResponder, PendingGetClosestType, SwarmDriver},
     error::{NetworkError, Result},
-    event::TerminateNodeReason,
+    interface::{LocalSwarmCmd, NetworkSwarmCmd, TerminateNodeReason},
     log_markers::Marker,
-    Addresses, MsgResponder, NetworkEvent, CLOSE_GROUP_SIZE,
+    Addresses, NetworkEvent, NodeIssue, SwarmLocalState, CLOSE_GROUP_SIZE,
 };
-use ant_evm::{PaymentQuote, QuotingMetrics};
-use ant_protocol::messages::ConnectionInfo;
+use ant_evm::PaymentQuote;
 use ant_protocol::{
-    messages::{Cmd, Request, Response},
+    messages::{Cmd, Request},
     storage::{DataTypes, RecordHeader, RecordKind, ValidationType},
     NetworkAddress, PrettyPrintRecordKey,
 };
 use libp2p::{
     kad::{
         store::{Error as StoreError, RecordStore},
-        KBucketDistance as Distance, Record, RecordKey,
+        KBucketDistance as Distance,
     },
     swarm::dial_opts::{DialOpts, PeerCondition},
     Multiaddr, PeerId,
 };
 use std::time::Instant;
-use std::{
-    collections::{BTreeMap, HashMap},
-    fmt::Debug,
-    time::Duration,
-};
+use std::{collections::BTreeMap, time::Duration};
 use tokio::sync::oneshot;
 use xor_name::XorName;
 
@@ -44,334 +39,6 @@ const REPLICATION_TIMEOUT: Duration = Duration::from_secs(45);
 
 // Throttles replication to at most once every 30 seconds
 const MIN_REPLICATION_INTERVAL_S: Duration = Duration::from_secs(30);
-
-#[derive(Debug, Eq, PartialEq, Clone)]
-pub(crate) enum NodeIssue {
-    /// Some connections might be considered to be critical and should be tracked.
-    ConnectionIssue,
-    /// Data Replication failed
-    ReplicationFailure,
-    /// Close nodes have reported this peer as bad
-    #[allow(dead_code)]
-    CloseNodesShunning,
-    /// Provided a bad quote
-    BadQuoting,
-    /// Peer failed to pass the chunk proof verification
-    FailedChunkProofCheck,
-}
-
-impl std::fmt::Display for NodeIssue {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            NodeIssue::ConnectionIssue => write!(f, "CriticalConnectionIssue"),
-            NodeIssue::ReplicationFailure => write!(f, "ReplicationFailure"),
-            NodeIssue::CloseNodesShunning => write!(f, "CloseNodesShunning"),
-            NodeIssue::BadQuoting => write!(f, "BadQuoting"),
-            NodeIssue::FailedChunkProofCheck => write!(f, "FailedChunkProofCheck"),
-        }
-    }
-}
-
-/// Commands to send to the Swarm
-pub(crate) enum LocalSwarmCmd {
-    /// Get a list of all peers in local RT, with correspondent Multiaddr info attached as well.
-    GetPeersWithMultiaddr {
-        sender: oneshot::Sender<Vec<(PeerId, Vec<Multiaddr>)>>,
-    },
-    /// Get a map where each key is the ilog2 distance of that Kbucket
-    /// and each value is a vector of peers in that bucket.
-    GetKBuckets {
-        sender: oneshot::Sender<BTreeMap<u32, Vec<PeerId>>>,
-    },
-    // Get K closest peers to target from the local RoutingTable, self is included
-    GetKCloseLocalPeersToTarget {
-        key: NetworkAddress,
-        sender: oneshot::Sender<Vec<(PeerId, Addresses)>>,
-    },
-    GetSwarmLocalState(oneshot::Sender<SwarmLocalState>),
-    /// Check if the local RecordStore contains the provided key
-    RecordStoreHasKey {
-        key: RecordKey,
-        sender: oneshot::Sender<bool>,
-    },
-    /// Get the Addresses of all the Records held locally
-    GetAllLocalRecordAddresses {
-        sender: oneshot::Sender<HashMap<NetworkAddress, ValidationType>>,
-    },
-    /// Get data from the local RecordStore
-    GetLocalRecord {
-        key: RecordKey,
-        sender: oneshot::Sender<Option<Record>>,
-    },
-    /// GetLocalQuotingMetrics for this node
-    /// Returns the quoting metrics and whether the record at `key` is already stored locally
-    GetLocalQuotingMetrics {
-        key: RecordKey,
-        data_type: u32,
-        data_size: usize,
-        sender: oneshot::Sender<(QuotingMetrics, bool)>,
-    },
-    /// Notify the node received a payment.
-    PaymentReceived,
-    /// Put record to the local RecordStore
-    PutLocalRecord {
-        record: Record,
-        is_client_put: bool,
-    },
-    /// Remove a local record from the RecordStore
-    /// Typically because the write failed
-    RemoveFailedLocalRecord {
-        key: RecordKey,
-    },
-    /// Add a local record to the RecordStore's HashSet of stored records
-    /// This should be done after the record has been stored to disk
-    AddLocalRecordAsStored {
-        key: RecordKey,
-        record_type: ValidationType,
-        data_type: DataTypes,
-    },
-    /// Add a peer to the blocklist
-    AddPeerToBlockList {
-        peer_id: PeerId,
-    },
-    /// Notify whether peer is in trouble
-    RecordNodeIssue {
-        peer_id: PeerId,
-        issue: NodeIssue,
-    },
-    // Whether peer is considered as `in trouble` by self
-    IsPeerShunned {
-        target: NetworkAddress,
-        sender: oneshot::Sender<bool>,
-    },
-    // Quote verification agaisnt historical collected quotes
-    QuoteVerification {
-        quotes: Vec<(PeerId, PaymentQuote)>,
-    },
-    // Notify a fetch completion
-    FetchCompleted((RecordKey, ValidationType)),
-    /// Triggers interval repliation
-    /// NOTE: This does result in outgoing messages, but is produced locally
-    TriggerIntervalReplication,
-    /// Triggers unrelevant record cleanup
-    TriggerIrrelevantRecordCleanup,
-    /// Send peer scores (collected from storage challenge) to replication_fetcher
-    NotifyPeerScores {
-        peer_scores: Vec<(PeerId, bool)>,
-    },
-    /// Add fresh replicate records into replication_fetcher
-    AddFreshReplicateRecords {
-        holder: NetworkAddress,
-        keys: Vec<(NetworkAddress, ValidationType)>,
-    },
-    /// Notify a fetched peer's version
-    NotifyPeerVersion {
-        peer: PeerId,
-        version: String,
-    },
-    /// Get responsible distance range.
-    GetNetworkDensity {
-        sender: oneshot::Sender<Option<Distance>>,
-    },
-    /// Remove peer from the routing table
-    RemovePeer {
-        peer: PeerId,
-    },
-}
-
-/// Commands to send to the Swarm
-pub(crate) enum NetworkSwarmCmd {
-    // Get closest peers from the network
-    GetClosestPeersToAddressFromNetwork {
-        key: NetworkAddress,
-        sender: oneshot::Sender<Vec<(PeerId, Addresses)>>,
-    },
-
-    // Send Request to the PeerId.
-    SendRequest {
-        req: Request,
-        peer: PeerId,
-        /// If the address is provided, we will try to perform a dial before sending the request.
-        addrs: Option<Addresses>,
-
-        // If a `sender` is provided, the requesting node will await for a `Response` from the
-        // Peer. The result is then returned at the call site.
-        //
-        // If a `sender` is not provided, the requesting node will not wait for the Peer's
-        // response. Instead we trigger a `NetworkEvent::ResponseReceived` which calls the common
-        // `response_handler`
-        #[allow(clippy::type_complexity)]
-        sender: Option<oneshot::Sender<Result<(Response, Option<ConnectionInfo>)>>>,
-    },
-    SendResponse {
-        resp: Response,
-        channel: MsgResponder,
-    },
-    // Attempt to dial specific peer.
-    DialPeer {
-        peer: PeerId,
-        addrs: Addresses,
-    },
-}
-
-/// Debug impl for LocalSwarmCmd to avoid printing full Record, instead only RecodKey
-/// and RecordKind are printed.
-impl Debug for LocalSwarmCmd {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            LocalSwarmCmd::PutLocalRecord {
-                record,
-                is_client_put,
-            } => {
-                write!(
-                    f,
-                    "LocalSwarmCmd::PutLocalRecord {{ key: {:?}, is_client_put: {is_client_put:?} }}",
-                    PrettyPrintRecordKey::from(&record.key)
-                )
-            }
-            LocalSwarmCmd::RemoveFailedLocalRecord { key } => {
-                write!(
-                    f,
-                    "LocalSwarmCmd::RemoveFailedLocalRecord {{ key: {:?} }}",
-                    PrettyPrintRecordKey::from(key)
-                )
-            }
-            LocalSwarmCmd::AddLocalRecordAsStored {
-                key,
-                record_type,
-                data_type,
-            } => {
-                write!(
-                    f,
-                    "LocalSwarmCmd::AddLocalRecordAsStored {{ key: {:?}, record_type: {record_type:?}, data_type: {data_type:?} }}",
-                    PrettyPrintRecordKey::from(key)
-                )
-            }
-            LocalSwarmCmd::GetKCloseLocalPeersToTarget { key, .. } => {
-                write!(
-                    f,
-                    "LocalSwarmCmd::GetKCloseLocalPeersToTarget {{ key: {key:?} }}"
-                )
-            }
-            LocalSwarmCmd::GetLocalQuotingMetrics { .. } => {
-                write!(f, "LocalSwarmCmd::GetLocalQuotingMetrics")
-            }
-            LocalSwarmCmd::PaymentReceived => {
-                write!(f, "LocalSwarmCmd::PaymentReceived")
-            }
-            LocalSwarmCmd::GetLocalRecord { key, .. } => {
-                write!(
-                    f,
-                    "LocalSwarmCmd::GetLocalRecord {{ key: {:?} }}",
-                    PrettyPrintRecordKey::from(key)
-                )
-            }
-            LocalSwarmCmd::GetAllLocalRecordAddresses { .. } => {
-                write!(f, "LocalSwarmCmd::GetAllLocalRecordAddresses")
-            }
-            LocalSwarmCmd::GetPeersWithMultiaddr { .. } => {
-                write!(f, "LocalSwarmCmd::GetPeersWithMultiaddr")
-            }
-            LocalSwarmCmd::GetKBuckets { .. } => {
-                write!(f, "LocalSwarmCmd::GetKBuckets")
-            }
-            LocalSwarmCmd::GetSwarmLocalState { .. } => {
-                write!(f, "LocalSwarmCmd::GetSwarmLocalState")
-            }
-            LocalSwarmCmd::RecordStoreHasKey { key, .. } => {
-                write!(
-                    f,
-                    "LocalSwarmCmd::RecordStoreHasKey {:?}",
-                    PrettyPrintRecordKey::from(key)
-                )
-            }
-            LocalSwarmCmd::AddPeerToBlockList { peer_id } => {
-                write!(f, "LocalSwarmCmd::AddPeerToBlockList {peer_id:?}")
-            }
-            LocalSwarmCmd::RecordNodeIssue { peer_id, issue } => {
-                write!(
-                    f,
-                    "LocalSwarmCmd::SendNodeStatus peer {peer_id:?}, issue: {issue:?}"
-                )
-            }
-            LocalSwarmCmd::IsPeerShunned { target, .. } => {
-                write!(f, "LocalSwarmCmd::IsPeerInTrouble target: {target:?}")
-            }
-            LocalSwarmCmd::QuoteVerification { quotes } => {
-                write!(
-                    f,
-                    "LocalSwarmCmd::QuoteVerification of {} quotes",
-                    quotes.len()
-                )
-            }
-            LocalSwarmCmd::FetchCompleted((key, record_type)) => {
-                write!(
-                    f,
-                    "LocalSwarmCmd::FetchCompleted({record_type:?} : {:?})",
-                    PrettyPrintRecordKey::from(key)
-                )
-            }
-            LocalSwarmCmd::TriggerIntervalReplication => {
-                write!(f, "LocalSwarmCmd::TriggerIntervalReplication")
-            }
-            LocalSwarmCmd::TriggerIrrelevantRecordCleanup => {
-                write!(f, "LocalSwarmCmd::TriggerUnrelevantRecordCleanup")
-            }
-            LocalSwarmCmd::NotifyPeerScores { peer_scores } => {
-                write!(f, "LocalSwarmCmd::NotifyPeerScores({peer_scores:?})")
-            }
-            LocalSwarmCmd::AddFreshReplicateRecords { holder, keys } => {
-                write!(
-                    f,
-                    "LocalSwarmCmd::AddFreshReplicateRecords({holder:?}, {keys:?})"
-                )
-            }
-            LocalSwarmCmd::NotifyPeerVersion { peer, version } => {
-                write!(f, "LocalSwarmCmd::NotifyPeerVersion({peer:?}, {version:?})")
-            }
-            LocalSwarmCmd::GetNetworkDensity { .. } => {
-                write!(f, "LocalSwarmCmd::GetNetworkDensity")
-            }
-            LocalSwarmCmd::RemovePeer { peer } => {
-                write!(f, "LocalSwarmCmd::RemovePeer({peer:?})")
-            }
-        }
-    }
-}
-
-/// Debug impl for NetworkSwarmCmd to avoid printing full Record, instead only RecodKey
-/// and RecordKind are printed.
-impl Debug for NetworkSwarmCmd {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            NetworkSwarmCmd::GetClosestPeersToAddressFromNetwork { key, .. } => {
-                write!(f, "NetworkSwarmCmd::GetClosestPeers {{ key: {key:?} }}")
-            }
-            NetworkSwarmCmd::SendResponse { resp, .. } => {
-                write!(f, "NetworkSwarmCmd::SendResponse resp: {resp:?}")
-            }
-            NetworkSwarmCmd::SendRequest { req, peer, .. } => {
-                write!(
-                    f,
-                    "NetworkSwarmCmd::SendRequest req: {req:?}, peer: {peer:?}"
-                )
-            }
-            NetworkSwarmCmd::DialPeer { peer, .. } => {
-                write!(f, "NetworkSwarmCmd::DialPeer peer: {peer:?}")
-            }
-        }
-    }
-}
-/// Snapshot of information kept in the Swarm's local state
-#[derive(Debug, Clone)]
-pub struct SwarmLocalState {
-    /// List of peers that we have an established connection with.
-    pub connected_peers: Vec<PeerId>,
-    /// The number of peers in the routing table
-    pub peers_in_routing_table: usize,
-    /// List of addresses the node is currently listening on
-    pub listeners: Vec<Multiaddr>,
-}
 
 impl SwarmDriver {
     pub(crate) fn handle_network_cmd(&mut self, cmd: NetworkSwarmCmd) -> Result<(), NetworkError> {
