+++ ant-node/src/networking/driver/event/kad.rs
@@ -7,11 +7,13 @@
 // permissions and limitations relating to use of the SAFE Network Software.
 
 use crate::networking::{
-    driver::PendingGetClosestType, Addresses, NetworkError, Result, SwarmDriver, CLOSE_GROUP_SIZE,
+    driver::PendingGetClosestType, Addresses, NetworkError, Result, CLOSE_GROUP_SIZE,
 };
 use libp2p::kad::{self, GetClosestPeersError, InboundRequest, QueryResult, K_VALUE};
 use std::collections::hash_map::Entry;
 use std::time::Instant;
+
+use super::SwarmDriver;
 
 impl SwarmDriver {
     pub(super) fn handle_kad_event(&mut self, kad_event: libp2p::kad::Event) -> Result<()> {
